refactor pack into unpack logic
