+++ src/packages/end-to-end/.gitignore
@@ -6,3 +6,5 @@ node_modules
 /test-results/
 /playwright-report/
 /playwright/.cache/
+
+local_modules
