+++ src/packages/end-to-end/package.json
@@ -5,9 +5,9 @@
 	"license": "MIT",
 	"scripts": {
 		"test-build": ". ./scripts/test-build.sh",
-		"import-database-sqlite": "tsx ./scripts/import-sqlite.ts",
+		"import-database-sqlite": "pnpm unpack && tsx ./scripts/import-sqlite.ts",
 		"import-database-postgres": ". ./scripts/import-postgres.sh",
-		"import-database-mysql": ". ./scripts/import-mysql.sh",
+		"import-database-mysql": "pnpm unpack && . ./scripts/import-mysql.sh",
 		"seed-postgres:reseed": "export DATABASE_RESEED=true && node ./scripts/seed-postgres.js",
 		"seed-postgres": "node ./scripts/seed-postgres.js",
 		"start-server": "cd app && pnpm start",
@@ -15,6 +15,7 @@
 		"dev-sqlite": "pnpm import-database-sqlite && pnpm start-server",
 		"dev-postgres": "createdb gw && pnpm seed-postgres && pnpm import-database-postgres  && pnpm start-server",
 		"dev-mysql": "pnpm import-database-mysql && pnpm start-server",
+		"unpack": "tsx ./scripts/unpack.ts",
 		"test-auth": "jest --runInBand --forceExit -- api/auth",
 		"test-mysql": "export DATABASE=mysql && jest --runInBand -- api/mysql",
 		"test-postgres": "export DATABASE=postgres && jest --runInBand -- api/postgres",
