+++ src/packages/end-to-end/scripts/import-mysql.sh
@@ -1,8 +1,5 @@
 rm -rf ./app 
-cd ../../
-pnpm publish:local
-cd packages/end-to-end
-node ../cli/bin init --name=app --backend=mysql --useVersion=\"local\" 
+node ./local_modules/graphweaver/bin init --name=app --backend=mysql --useVersion=\"local\" 
 cd app 
 pnpm i --ignore-workspace --no-lockfile
 mkdir databases
@@ -10,4 +7,4 @@ cp ../databases/mysql.sql databases/mysql.sql
 mysql -h$DATABASE_HOST -u$DATABASE_USERNAME --password=$DATABASE_PASSWORD -e "DROP DATABASE IF EXISTS Chinook;"
 mysql -h$DATABASE_HOST -u$DATABASE_USERNAME --password=$DATABASE_PASSWORD -e "CREATE DATABASE Chinook;"
 mysql -h$DATABASE_HOST -u$DATABASE_USERNAME --password=$DATABASE_PASSWORD Chinook < databases/mysql.sql
-node ../../cli/bin import mysql --database=Chinook --user=$DATABASE_USERNAME --password=$DATABASE_PASSWORD --host=$DATABASE_HOST --port=3306 --o
+pnpm run import mysql --database=Chinook --user=$DATABASE_USERNAME --password=$DATABASE_PASSWORD --host=$DATABASE_HOST --port=3306 --o
